Anchor

+++ contracts/interfaces/ICoFiXPool.sol
@@ -30,14 +30,14 @@ interface ICoFiXPool {
     /// @dev Set configuration
     /// @param theta Trade fee rate, ten thousand points system. 20
     /// @param impactCostVOL Impact cost threshold
-    /// @param nt Each unit token (in the case of binary pools, eth) is used for the standard ore output, 1e9 based
-    function setConfig(uint16 theta, uint16 impactCostVOL, uint56 nt) external;
+    /// @param nt Each unit token (in the case of binary pools, eth) is used for the standard ore output, 1e18 based
+    function setConfig(uint16 theta, uint96 impactCostVOL, uint96 nt) external;
 
     /// @dev Get configuration
     /// @return theta Trade fee rate, ten thousand points system. 20
     /// @return impactCostVOL Impact cost threshold
-    /// @return nt Each unit token (in the case of binary pools, eth) is used for the standard ore output, 1e9 based
-    function getConfig() external view returns (uint16 theta, uint16 impactCostVOL, uint56 nt);
+    /// @return nt Each unit token (in the case of binary pools, eth) is used for the standard ore output, 1e18 based
+    function getConfig() external view returns (uint16 theta, uint96 impactCostVOL, uint96 nt);
 
     /// @dev Add liquidity and mint xtoken
     /// @param token Target token address
